requirements.txt
pypresence
pillow
pystray
flask
requests
psutil
pywin32
nest_asyncio 